Added files via upload
SAMPLE PROBLEM B ARE MY SUMS RIGHT?
In this problem, you will be presented with a series of lines each containing a sum (addition,
subtraction or multiplication only). All you have to do is mark them - is the answer correct?
Input
Input consists of a series of lines, terminated by a line containing just a # - this line should
not be processed. Each line contains an integer, an operator (+, - or *), a second integer,
an equals sign and a third integer, all separated by single spaces. If the operator is applied
to the first two integers, is the result the third integer?
Output
Output consists of a series of lines, one for each of the lines in the input. Each line consists
of the word Correct if the third integer is the result of applying the operator to the first two
integers, or Wrong otherwise.
Sample Input
1 + 2 = 3
2 * 2 = 5
6 ‐ 2 = 4
#
Output for Sample Input
Correct
Wrong
Correct