Added files via upload
SAMPLE PROBLEM C UNTIED AIRLINES
An airline called Untied, has had a series of bad press articles. The flying public are not
impressed. So Untied has decided to award upgrades on future flights if a passenger
suffers certain elements of poor service.
The poor service includes things like lost luggage, the stewardess swiping a passenger in
the face as she reaches past them, or cancelling flights at short notice and not offering
alternative flights in time for connections.
These elements of poor service are encoded and each is awarded points. If a passenger
collects 200 such points on one flight, then their next flight is automatically upgraded. The
poor service codes and associated points are shown in the table below.
Your job is to tell Untied for each flight, how many upgrades are awarded.
Input
Input consists of data for a single flight which begins with a flight number. The flight number
starts with UT, in upper case, followed by a sequence of up to 4 digits (eg: UT525)